upgraded entire project to typescript and re-implimented webpack. New cleaner sturcutrie and more understandable code. Comment to come.
