+++ .gitignore
@@ -1,20 +1,8 @@
 # dependencies
 /node_modules
 
-# testing
-/coverage
-
-# production
-/build
+# builds
 /dist
 
-# misc
-.DS_Store
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
-
+# npm
 npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
