+++ package.json
@@ -12,53 +12,48 @@
     "Reece Harrison <reece@harrison.io>"
   ],
   "license": "MIT",
-  "private": true,
   "scripts": {
-    "start": "webpack-dev-server -w --config tools/webpack/webpack.common.js",
+    "build": "npm run clean-dist && webpack -p --config=webpack/prod.js",
+    "clean-dist": "rm -f -r -d dist",
+    "lint": "npm run lint:ts && npm run lint:sass",
+    "lint:ts": "tslint './src/**/*.ts*' --format stylish --force",
+    "start": "npm run start-dev",
+    "start-dev": "webpack-dev-server --config=webpack/dev.js",
+    "start-prod": "npm run build",
     "deploy": "gh-pages -d build"
   },
-  "dependencies": {
-    "antd": "^2.10.4",
-    "bzrest": "^1.0.4-alpha",
-    "json-loader": "^0.5.7",
-    "lodash": "^4.17.4",
-    "prop-types": "^15.5.10",
-    "raw-loader": "^0.5.1",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
-    "sass-loader": "^6.0.6",
-    "source-map-loader": "^0.2.3"
-  },
   "devDependencies": {
-    "awesome-typescript-loader": "^3.4.1",
-    "case-sensitive-paths-webpack-plugin": "1.1.4",
-    "chalk": "1.1.3",
-    "css-loader": "^0.28.1",
-    "dotenv": "4.0.0",
-    "extract-text-webpack-plugin": "2.1.0",
-    "file-loader": "0.11.1",
-    "fs-extra": "3.0.1",
-    "gh-pages": "^1.0.0",
-    "html-webpack-plugin": "2.28.0",
+    "@types/node": "^8.0.53",
+    "awesome-typescript-loader": "^3.4.0",
+    "css-loader": "^0.28.7",
+    "express": "^4.16.2",
+    "file-loader": "^1.1.5",
+    "html-webpack-plugin": "^2.30.1",
+    "image-webpack-loader": "^3.4.2",
     "node-sass": "^4.7.2",
-    "object-assign": "4.1.1",
-    "postcss-cssnext": "^2.11.0",
-    "postcss-flexbugs-fixes": "3.0.0",
-    "postcss-loader": "2.0.5",
-    "promise": "7.1.1",
-    "react-dev-utils": "^3.0.0",
-    "react-error-overlay": "^1.0.7",
-    "style-loader": "0.17.0",
-    "sw-precache-webpack-plugin": "0.9.1",
-    "ts-node": "^3.3.0",
-    "tslib": "^1.8.0",
+    "postcss-loader": "^2.0.9",
+    "react": "^16.2.0",
+    "react-addons-test-utils": "^15.6.2",
+    "react-dom": "^16.2.0",
+    "react-hot-loader": "^3.1.3",
+    "sass-loader": "^6.0.6",
+    "style-loader": "^0.19.0",
+    "stylelint": "^8.3.1",
+    "stylelint-config-standard": "^18.0.0",
+    "stylelint-webpack-plugin": "^0.9.0",
     "tslint": "^5.8.0",
     "typescript": "^2.6.2",
-    "url-loader": "0.5.8",
-    "webpack": "2.6.1",
-    "webpack-dev-server": "2.4.5",
-    "webpack-manifest-plugin": "1.1.0",
+    "uglifyjs-webpack-plugin": "^1.1.1",
+    "webpack": "^3.8.1",
+    "webpack-dev-middleware": "^1.12.2",
+    "webpack-dev-server": "^2.9.5",
     "webpack-merge": "^4.1.1",
-    "whatwg-fetch": "2.0.3"
-  }
+    "@types/react": "^16.0.25"
+  },
+  "dependencies": {
+    "antd": "^2.13.10",
+    "bzrest": "^1.0.4-alpha",
+    "lodash": "^4.17.4"
+  },
+  "postcss": {}
 }
