+++ src/css/index.scss
@@ -1,4 +1,4 @@
-/* @import url('https://fonts.googleapis.com/css?family=Roboto');
+@import url('https://fonts.googleapis.com/css?family=Roboto');
 @import url('https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css');
 
 :root {
@@ -142,4 +142,4 @@ body, html {
 /*.ant-card-extra {
     right: 10px;
     top: 10px;
-}*/ */
+}*/
