+++ src/index.html
@@ -1,14 +1,11 @@
-<!-- <!doctype html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, user-scalable=no, minimum-scale=1.0, maximum-scale=1.0">
-    <link rel="shortcut icon" href="favicon.ico">
-
-    <title>Moz Bugs</title>
-  </head>
-  <body>
-    <script src="dist/bundle.js>"></script>
-    <div id="root"></div>
-  </body>
-</html>  -->
+<!DOCTYPE html>
+<html lang="en" dir="ltr">
+    <head>
+        <meta charset="UTF-8"/>
+        <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+        <title>Hello React!</title>
+    </head>
+    <body>
+        <div id="root"></div>       
+    </body>
+</html>
