+++ src/index.tsx
@@ -1,9 +1,9 @@
-import React from 'react'
+import * as React from 'react'
 import { render } from 'react-dom'
 
-import App from './app'
+import { App } from './app'
 
-import './css/index.css'
+// import './css/index.css'
 
 render(
   <App />,
