--- tsconfg.json
@@ -1,44 +1,0 @@
-{
-    "compilerOptions": {
-      "module": "es2015",
-      "target": "es2015",
-      "jsx": "react",
-      "rootDir": "src",
-      "outDir": "dist",
-      "allowSyntheticDefaultImports": true,
-      "noImplicitAny": true,
-      "sourceMap": true,
-      "experimentalDecorators": true,
-      "preserveConstEnums": true,
-      "allowJs": true,
-      "noUnusedLocals": true,
-      "noUnusedParameters": true,
-      "noImplicitReturns": true,
-      "moduleResolution": "node",
-      "baseUrl": "src"
-    },
-    "filesGlob": [
-      "typings/index.d.ts",
-      "src/**/*.ts",
-      "src/**/*.tsx",
-      "node_modules/typescript/lib/lib.es6.d.ts"
-    ],
-    "exclude": [
-      "build",
-      "dist",
-      "node_modules",
-      "__tests__"
-    ],
-    "types": [
-      "node",
-      "webpack",
-      "react",
-      "react-dom",
-      "reactstrap"
-    ],
-    "include": [
-      "**/*.ts",
-      "**/*.tsx"
-    ],
-    "compileOnSave": false
-  }
