+++ webpack/common.js
@@ -1,0 +1,46 @@
+// shared config (dev and prod)
+const { resolve } = require('path');
+const { CheckerPlugin } = require('awesome-typescript-loader');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+
+module.exports = {
+  resolve: {
+    extensions: ['.ts', '.tsx', '.js', '.jsx'],
+  },
+  context: resolve(__dirname, '../src'),
+  module: {
+    rules: [
+      {
+        test: /\.tsx?$/,
+        use: 'awesome-typescript-loader',
+      },
+      {
+        test: /\.css$/,
+        use: ['style-loader', { loader: 'css-loader', options: { importLoaders: 1 } }, 'postcss-loader',],
+      },
+      {
+        test: /\.scss$/,
+        loaders: [
+          'style-loader',
+          { loader: 'css-loader', options: { importLoaders: 1 } },
+          'postcss-loader',
+          'sass-loader',
+        ],
+      },
+      {
+        test: /\.(jpe?g|png|gif|svg)$/i,
+        loaders: [
+          'file-loader?hash=sha512&digest=hex&name=img/[hash].[ext]',
+          'image-webpack-loader?bypassOnDebug&optipng.optimizationLevel=7&gifsicle.interlaced=false',
+        ],
+      },
+    ],
+  },
+  plugins: [
+    new CheckerPlugin(),
+    new HtmlWebpackPlugin({ template: 'index.html', }),
+  ],
+  performance: {
+    hints: true,
+  },
+};
